chore: moved routing-related files to a folder
