+++ packages/frontend/src/app/main/main.component.spec.ts
@@ -23,7 +23,7 @@ import {MainComponent} from './main.component';
 import {Component, Output, EventEmitter, Input} from '@angular/core';
 import {Search} from '../services/search/interfaces';
 import {SearchService} from '../services/search/search.service';
-import {AppRoutingModule} from '../app-routing.module';
+import {AppRoutingModule} from '../routing/app-routing.module';
 import {of} from 'rxjs';
 import {By} from '@angular/platform-browser';
 import {Location} from '@angular/common';
